refactor: rename and isolate legacy Panel component for migration
- Renamed `Panel` to `PanelOld` and created a new file to facilitate the transition to a unified `Panel` wrapper.
- Updated imports across various components to reference the new `PanelOld` while maintaining functionality.
- Completed initial checklist items for the panel refactor, ensuring legacy usage is documented and the repository compiles correctly.
