+++ app/routes/a7k9m2x5p8w1n4q6r3y8b5t1/tournaments/tournaments._index.tsx
@@ -6,7 +6,7 @@ import { redirect, useLoaderData, useRevalidator, useSubmit } from 'react-router
 import { Box, Flex, Grid, Heading, Text } from '@radix-ui/themes'
 
 import { DeleteIcon, TrophyIcon } from '~/components/icons'
-import { Panel } from '~/components/Panel'
+import { PanelOld as Panel } from '~/components/PanelOld'
 import {
   datatableActionButtonVariants,
   datatableCellTextVariants,
